fixe problem of last commit, complete method in Classifier for execute query and make connection with database
